pyproject.toml
[tool.poetry]
name = "textures-ambientcg"
version = "2023.07.19"
description = "Fetch ambientcg.com textures"
authors = ["Your Name <you@example.com>"]
license = "WTFPL"
readme = "README.md"
# packages = [{include = "textures_ambientcg"}]

[tool.poetry.dependencies]
python = "^3.11"
ipdb = "^0.13.13"
requests = "^2.31.0"
requests-cache = "^1.1.0"
Pillow = "^10.0.0"
Jinja2 = "^3.1.2"
tomlkit = "^0.11.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
